Updated Lab 4
I made the changes that we discussed in class today to page 5 of lab 4 and page 8 of lab 4. Page 5 was missing the print statements which I was able to add while page 8's code did not compile, so I removed the comment from the variable, and now it is able to run properly.